requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
Pillow>=8.3.0

# Federated Learning
flwr>=1.4.0

# Data processing
scikit-learn>=1.0.0
pandas>=1.3.0

# Visualization and logging
matplotlib>=3.5.0
tensorboard>=2.8.0
tqdm>=4.62.0

# Development and testing
pytest>=6.2.0
black>=21.0.0
flake8>=4.0.0

# Optional: For advanced models
# timm>=0.6.0  # For additional model architectures
# efficientnet-pytorch>=0.7.0  # Alternative EfficientNet implementation
